Moving mbedtls specific utility functions to seperate files
