+++ tests/tls_e2e/common/utility.cpp
@@ -3,12 +3,12 @@
 
 // clang-format off
 #include <openenclave/enclave.h>
-#include <openenclave/internal/raise.h>
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
 #include "tls_e2e_t.h"
 #include "utility.h"
+
 // clang-format on
 
 extern struct tls_control_args g_control_config;
@@ -63,256 +63,57 @@ done:
     return result;
 }
 
-// Compute the sha256 hash of given data.
-static int sha256(const uint8_t* data, size_t data_size, uint8_t sha256[32])
+// This is the identity validation callback. A TLS connecting party (client or
+// server) can verify the passed in "identity" information to decide whether to
+// accept an connection request
+oe_result_t enclave_identity_verifier(oe_identity_t* identity, void* arg)
 {
-    int ret = 0;
-    mbedtls_sha256_context ctx;
+    oe_result_t result = OE_VERIFY_FAILED;
 
-    mbedtls_sha256_init(&ctx);
+    (void)arg;
 
-    ret = mbedtls_sha256_starts_ret(&ctx, 0);
-    if (ret)
+    OE_TRACE_INFO("enclave_identity_verifier is called with parsed report:\n");
+    if (g_control_config.fail_enclave_identity_verifier_callback)
         goto done;
 
-    ret = mbedtls_sha256_update_ret(&ctx, data, data_size);
-    if (ret)
-        goto done;
-
-    ret = mbedtls_sha256_finish_ret(&ctx, sha256);
-    if (ret)
-        goto done;
-
-done:
-    mbedtls_sha256_free(&ctx);
-    return ret;
-}
-
-oe_result_t generate_certificate_and_pkey(
-    mbedtls_x509_crt* cert,
-    mbedtls_pk_context* private_key)
-{
-    oe_result_t result = OE_FAILURE;
-    uint8_t* output_cert = NULL;
-    size_t output_cert_size = 0;
-    uint8_t* private_key_buf = NULL;
-    size_t private_key_buf_size = 0;
-    uint8_t* public_key_buf = NULL;
-    size_t public_key_buf_size = 0;
-    int ret = 0;
-
-    result = generate_key_pair(
-        &public_key_buf,
-        &public_key_buf_size,
-        &private_key_buf,
-        &private_key_buf_size);
-    OE_CHECK_MSG(result, " failed with %s\n", oe_result_str(result));
-
-#if 0
-    OE_TRACE_INFO("private_key_buf_size:[%ld]\n", private_key_buf_size);
-#endif
-
-    OE_TRACE_INFO("public_key_buf_size:[%ld]\n", public_key_buf_size);
-    OE_TRACE_INFO("public key used:\n[%s]", public_key_buf);
-
-    result = oe_generate_attestation_certificate(
-        (const unsigned char*)"CN=Open Enclave SDK,O=OESDK TLS,C=US",
-        private_key_buf,
-        private_key_buf_size,
-        public_key_buf,
-        public_key_buf_size,
-        &output_cert,
-        &output_cert_size);
-    OE_CHECK_MSG(result, " failed with %s\n", oe_result_str(result));
-
-    // create mbedtls_x509_crt from output_cert
-    ret = mbedtls_x509_crt_parse_der(cert, output_cert, output_cert_size);
-    if (ret != 0)
-        OE_RAISE_MSG(OE_VERIFY_FAILED, " failed with ret = %d\n", ret);
-
-    // create mbedtls_pk_context from private key data
-    ret = mbedtls_pk_parse_key(
-        private_key,
-        (const unsigned char*)private_key_buf,
-        private_key_buf_size,
-        NULL,
-        0);
-    if (ret != 0)
-        OE_RAISE_MSG(OE_VERIFY_FAILED, " failed with ret = %d\n", ret);
-
-done:
-    oe_free_key(private_key_buf, private_key_buf_size, NULL, 0);
-    oe_free_key(public_key_buf, public_key_buf_size, NULL, 0);
-    oe_free_attestation_certificate(output_cert);
-
-    return result;
-}
-
-bool verify_mrsigner(
-    char* siging_public_key_buf,
-    size_t siging_public_key_buf_size,
-    uint8_t* signer_id_buf,
-    size_t signer_id_buf_size)
-{
-    mbedtls_pk_context ctx;
-    mbedtls_pk_type_t pk_type;
-    mbedtls_rsa_context* rsa_ctx = NULL;
-    uint8_t* modulus = NULL;
-    size_t modulus_size = 0;
-    int res = 0;
-    bool ret = false;
-    unsigned char* signer = NULL;
-
-    mbedtls_pk_init(&ctx);
-    signer = (unsigned char*)malloc(signer_id_buf_size);
-    if (signer == NULL)
-    {
-        OE_TRACE_ERROR("Out of memory\n");
-        goto done;
-    }
-
-    OE_TRACE_INFO("Verify connecting client's identity\n");
-    OE_TRACE_INFO(
-        "public key buffer size[%lu]\n", sizeof(siging_public_key_buf));
-    OE_TRACE_INFO("public key\n[%s]\n", siging_public_key_buf);
-
-    res = mbedtls_pk_parse_public_key(
-        &ctx,
-        (const unsigned char*)siging_public_key_buf,
-        siging_public_key_buf_size);
-    if (res != 0)
-    {
-        OE_TRACE_ERROR("mbedtls_pk_parse_public_key failed with %d\n", res);
-        goto done;
-    }
-    OE_TRACE_INFO(
-        "siging_public_key_buf_size=%ld\n", siging_public_key_buf_size);
-
-    pk_type = mbedtls_pk_get_type(&ctx);
-    if (pk_type != MBEDTLS_PK_RSA)
-    {
-        OE_TRACE_ERROR("mbedtls_pk_get_type had incorrect type: %d\n", res);
-        goto done;
-    }
-    OE_TRACE_INFO("This public sigining key is a rsa key \n");
-
-    rsa_ctx = mbedtls_pk_rsa(ctx);
-    modulus_size = mbedtls_rsa_get_len(rsa_ctx);
-    OE_TRACE_INFO("modulus_size = [%zu]\n", modulus_size);
-    modulus = (uint8_t*)malloc(modulus_size);
-    if (modulus == NULL)
+    // Check the enclave's security version
+    if (identity->security_version < 1)
     {
         OE_TRACE_ERROR(
-            "malloc for modulus failed with size %zu:\n", modulus_size);
+            "identity->security_version checking failed (%d)\n",
+            identity->security_version);
         goto done;
     }
 
-    res = mbedtls_rsa_export_raw(
-        rsa_ctx, modulus, modulus_size, NULL, 0, NULL, 0, NULL, 0, NULL, 0);
-    if (res != 0)
-    {
-        OE_TRACE_ERROR("mbedtls_rsa_export failed with %d\n", res);
-        goto done;
-    }
+    // Dump an enclave's unique ID, signer ID and Product ID. They are
+    // MRENCLAVE, MRSIGNER and ISVPRODID for SGX enclaves. In a real scenario,
+    // custom id checking should be done here
+    OE_TRACE_INFO("\nidentity->unique_id :\n");
+    for (int i = 0; i < OE_UNIQUE_ID_SIZE; i++)
+        OE_TRACE_INFO("0x%0x ", (uint8_t)identity->unique_id[i]);
 
-    // Reverse the modulus and compute sha256 on it.
-    for (size_t i = 0; i < modulus_size / 2; i++)
-    {
-        uint8_t tmp = modulus[i];
-        modulus[i] = modulus[modulus_size - 1 - i];
-        modulus[modulus_size - 1 - i] = tmp;
-    }
+    OE_TRACE_INFO("\nparsed_report->identity.signer_id :\n");
+    for (int i = 0; i < OE_SIGNER_ID_SIZE; i++)
+        OE_TRACE_INFO("0x%0x ", (uint8_t)identity->signer_id[i]);
 
-    // Calculate the MRSIGNER value which is the SHA256 hash of the
-    // little endian representation of the public key modulus. This value
-    // is populated by the signer_id sub-field of a parsed oe_report_t's
-    // identity field.
+    // On a real enclave product, this is the place to check again the enclave
+    // signing key by calling function like verify_mrsigner below. However, we
+    // are not siging test cases, so this checking will be skipped.
+    // tls_between_enclaves sample will have code show how to dothis checking if
+    // (!verify_mrsigner((char *)OTHER_ENCLAVE_PUBLIC_KEY,
+    //                     sizeof(OTHER_ENCLAVE_PUBLIC_KEY),
+    //                     identity->signer_id,
+    //                     sizeof(identity->signer_id)))
+    // {
+    //     OE_TRACE_ERROR("failed:mrsigner not equal!\n");
+    //     goto done;
+    // }
 
-    OE_TRACE_INFO("modulus_size=%ld\n", modulus_size);
-    if (sha256(modulus, modulus_size, signer) != 0)
-    {
-        OE_TRACE_ERROR("sha256 failed\n");
-        goto done;
-    }
+    OE_TRACE_INFO("\nidentity->product_id :\n");
+    for (int i = 0; i < OE_PRODUCT_ID_SIZE; i++)
+        OE_TRACE_INFO("0x%0x ", (uint8_t)identity->product_id[i]);
 
-    if (memcmp(signer, signer_id_buf, signer_id_buf_size) != 0)
-    {
-        OE_TRACE_ERROR("mrsigner is not equal!\n");
-        for (size_t i = 0; i < signer_id_buf_size; i++)
-        {
-            OE_TRACE_ERROR(
-                "0x%x - 0x%x\n", (uint8_t)signer[i], (uint8_t)signer_id_buf[i]);
-        }
-        goto done;
-    }
-    ret = true;
+    result = OE_OK;
 done:
-    if (signer)
-        free(signer);
-
-    if (modulus != NULL)
-        free(modulus);
-
-    mbedtls_pk_free(&ctx);
-    return ret;
-}
-
-// If set, the verify callback is called for each certificate in the chain.
-// The verification callback is supposed to return 0 on success. Otherwise, the
-// verification failed.
-// The function should return 0 for anything (including invalid
-// certificates) other than fatal error, as a non-zero return
-// code immediately aborts the verification process.
-// For fatal errors, a specific error code should be used
-// (different from MBEDTLS_ERR_X509_CERT_VERIFY_FAILED which
-// should not be returned at this point), or MBEDTLS_ERR_X509_FATAL_ERROR
-// can be used if no better code is available.
-//
-int cert_verify_callback(
-    void* data,
-    mbedtls_x509_crt* crt,
-    int depth,
-    uint32_t* flags)
-{
-    oe_result_t result = OE_FAILURE;
-    int ret = MBEDTLS_ERR_X509_CERT_VERIFY_FAILED;
-    unsigned char* cert_buf = NULL;
-    size_t cert_size = 0;
-
-    (void)data;
-
-    *flags = (uint32_t)MBEDTLS_ERR_X509_CERT_VERIFY_FAILED;
-    if (g_control_config.fail_cert_verify_callback)
-    {
-        OE_TRACE_INFO(
-            "Purposely returns failure from server's cert_verify_callback()\n");
-        goto done;
-    }
-
-    cert_buf = crt->raw.p;
-    cert_size = crt->raw.len;
-
-    OE_TRACE_INFO("cert_verify_callback with depth = %d\n", depth);
-    OE_TRACE_INFO("crt->version = %d\n", crt->version);
-    OE_TRACE_INFO("cert_size = %zu\n", cert_size);
-
-    if (cert_size <= 0)
-        goto done;
-
-    OE_TRACE_INFO("Calling oe_verify_attestation_certificate\n");
-    if (g_control_config.fail_oe_verify_attestation_certificate)
-        goto done;
-
-    result = oe_verify_attestation_certificate(
-        cert_buf, cert_size, enclave_identity_verifier, NULL);
-    OE_CHECK_MSG(
-        result,
-        "oe_verify_attestation_certificate failed with result = %s\n",
-        oe_result_str(result));
-
-    OE_TRACE_INFO("\nReturned from oe_verify_attestation_certificate\n");
-    ret = 0;
-    *flags = 0;
-done:
-    return ret;
+    return result;
 }
