+++ tests/tls_e2e/common/utility.h
@@ -2,34 +2,23 @@
 // Licensed under the MIT License.
 
 // clang-format off
-#include <mbedtls/pk.h>
-#include <mbedtls/rsa.h>
-#include <mbedtls/sha256.h>
-#include <mbedtls/x509_crt.h>
-#include <mbedtls/platform.h>
+
 // clang-format on
 
 #include <openenclave/internal/print.h>
+#include <openenclave/internal/raise.h>
+#include <openenclave/internal/report.h>
+#include <openenclave/enclave.h>
 
-oe_result_t generate_certificate_and_pkey(
-    mbedtls_x509_crt* cert,
-    mbedtls_pk_context* private_key);
-int cert_verify_callback(
-    void* data,
-    mbedtls_x509_crt* crt,
-    int depth,
-    uint32_t* flags);
 oe_result_t enclave_identity_verifier(oe_identity_t* identity, void* arg);
 
-bool verify_mrsigner(
-    char* siging_public_key_buf,
-    size_t siging_public_key_buf_size,
-    uint8_t* signer_id_buf,
-    size_t signer_id_buf_size);
+oe_result_t generate_key_pair(
+    uint8_t** public_key,
+    size_t* public_key_size,
+    uint8_t** private_key,
+    size_t* private_key_size);
 
-// mbedtls debug levels: 0 No debug, 1 Error, 2 State change, 3 Informational, 4
-// Verbose
-#define DEBUG_LEVEL 1
+#define SERVER_IP "127.0.0.1"
 
 #define CLIENT_REQUEST_PAYLOAD_SIZE 18
 
@@ -42,3 +31,6 @@ bool verify_mrsigner(
     "A message from TLS server inside enclave\r\n"
 
 #define SERVER_RESPONSE_PAYLOAD_SIZE 194
+
+const unsigned char certificate_subject_name[] =
+    "CN=Open Enclave SDK,O=OESDK TLS,C=US";
